+++ lib/engineyard/cli.rb
@@ -363,12 +363,18 @@ WARNING: Interrupting again may prevent Engine Yard Cloud from recording this
       :desc => "Run command on the slave database servers"
     method_option :utilities, :type => :array, :lazy_default => true,
       :desc => "Run command on the utility servers with the given names. If no names are given, run on all utility servers."
+    method_option :shell, :type => :string, :lazy_default => 'bash', :aliases => %w(-s),
+      :desc => "Run command in a shell other than bash. Use --no-shell to run the command without a shell."
 
     def ssh(cmd=nil)
       environment = fetch_environment(options[:environment], options[:account])
       hosts = ssh_hosts(options, environment)
 
       raise NoCommandError.new if cmd.nil? and hosts.size != 1
+
+      if options[:shell] && cmd
+        cmd = Escape.shell_command([options[:shell],'-lc',cmd])
+      end
 
       exits = hosts.map do |host|
         system Escape.shell_command(['ssh', "#{environment.username}@#{host}", cmd].compact)